backend/pyproject.toml
[project]
name = "image-upload-demo"
version = "0.0.1"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "boto3==1.38.16",
    "botocore==1.38.16",
    "certifi==2025.4.26",
    "click==8.2.0",
    "dnspython==2.7.0",
    "email-validator==2.2.0",
    "fastapi==0.115.12",
    "fastapi-cli==0.0.7",
    "greenlet==3.2.2",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "idna==3.10",
    "jinja2==3.1.6",
    "jmespath==1.0.1",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "mdurl==0.1.2",
    "psycopg==3.2.8",
    "pydantic==2.11.4",
    "pydantic-core==2.33.2",
    "pygments==2.19.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "python-multipart==0.0.20",
    "pyyaml==6.0.2",
    "rich==14.0.0",
    "rich-toolkit==0.14.6",
    "s3transfer==0.12.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.40",
    "starlette==0.46.2",
    "typer==0.15.3",
    "typing-inspection==0.4.0",
    "typing-extensions==4.13.2",
    "urllib3==2.4.0",
    "uvicorn==0.34.2",
    "uvloop==0.21.0",
    "watchfiles==1.0.5",
    "websockets==15.0.1",
    "boto3-stubs[s3]>=1.38.16",
]

[tool.black]
line-length = 79

[tool.ruff]
line-length = 79

[tool.ruff.lint]
select = ["ALL"]

[tool.uv]
package = false

[tool.pyright]
reportUnknownMemberType = false

